requirements.txt
attrs==22.2.0
ConfigParser==7.0.0
ipython==8.12.3
ipywidgets==8.0.4
Jinja2==3.1.2
nltk==3.8.1
protobuf==5.27.2
pywin32==305
scikit_learn==1.5.1
tornado==6.2
python-docx==1.1.2